move UITableView extensions to own file add UITableView extensions tests refactor file names and headers to one style
